requirements.txt
requests
paho-mqtt
pika